Fix display of copyright in decompressor.

+++ cpp/Kanzi/io/CompressedOutputStream.cpp
@@ -455,7 +455,7 @@ T EncodingTask<T>::call() THROW
             if (postTransformLength < 0)
                 return EncodingTaskResult(_blockId, Error::ERR_WRITE_FILE, "Invalid transform size");
 
-            for (uint64 n = 0xFF; n < postTransformLength; n <<= 8)
+            for (uint64 n = 0xFF; n < uint64(postTransformLength); n <<= 8)
                 dataSize++;
 
             if (dataSize > 3)
